Day 06, part 1
Turns out we didn't need to abstract away the parsing since it changed
enough for part 2 for me to just write a new one for it.
So I rewrote the parsing for part 1 to be self contained in the part_1
function, except for the parts that are repeated, like the splitting
into numbers, and the stripping of spaces.